swagger ui initialized in routes folder
